+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes
 * New base class DSF for data stream Filters.
-* DSD_ScaleStream is now DSF_ScaleStream and DSD_ScaleStream is deprecated.
+* DSD_ScaleStream is now DSF_Scale and DSD_ScaleStream is deprecated.
+* New DSF: DSF_Downsample
 * The NAMESPACE is now managed using roxygen.
 
 # stream 1.5-1 (05/09/22)